+++ src/main/java/com/notenoughmail/kubejs_tfc/KubeJSTFC.java
@@ -68,7 +68,6 @@ import java.util.stream.Stream;
 
 import static com.notenoughmail.kubejs_tfc.util.implementation.commands.DataType.append;
 
-// TODO: 1.3.3 | Properly implement #replaceOutput/#replaceInput/#isOutput/#isInput for custom recipe components
 @SuppressWarnings("unused")
 @Mod(KubeJSTFC.MODID)
 public class KubeJSTFC {
@@ -76,7 +75,7 @@ public class KubeJSTFC {
     public static final String MOD_NAME = "KubeJS TFC";
     private static final Logger LOGGER = LogUtils.getLogger();
     public static final String MODID = "kubejs_tfc";
-    public static boolean debug, insertIntoConsole, deduplicateConsoleErrors;
+    public static boolean debug, insertIntoConsole, deduplicateConsoleErrors, asyncRhinoContext;
 
     private static Consumer<ImmutableMap.Builder<String, RegistryRock>> rockListeners = r -> {
         for (Rock rock : Rock.VALUES) {
@@ -99,6 +98,7 @@ public class KubeJSTFC {
         debug = props.debugInfo;
         insertIntoConsole = props.get("tfc/insertSelfTestsIntoConsole", true);
         deduplicateConsoleErrors = props.get("tfc/deduplicateConsoleErrors", true);
+        asyncRhinoContext = props.get("tfc/asyncRhinoContext", true);
 
         printConfig(KubeJSTFC::info);
     }
@@ -108,6 +108,7 @@ public class KubeJSTFC {
         info.accept("- Debug mode enabled: %s".formatted(debug));
         info.accept("- Self tests console insertion enabled: %s".formatted(insertIntoConsole));
         info.accept("- Self tests warnings deduplicated: %s".formatted(deduplicateConsoleErrors));
+        info.accept("- Async Rhino Context wrappers: %s".formatted(asyncRhinoContext));
     }
 
     public static void info(String message) {