+++ src/main/java/com/notenoughmail/kubejs_tfc/recipe/component/FluidIngredientComponent.java
@@ -3,6 +3,7 @@ package com.notenoughmail.kubejs_tfc.recipe.component;
 import com.google.gson.JsonElement;
 import com.notenoughmail.kubejs_tfc.util.helpers.IngredientHelpers;
 import dev.latvian.mods.kubejs.recipe.RecipeJS;
+import dev.latvian.mods.kubejs.recipe.ReplacementMatch;
 import dev.latvian.mods.kubejs.recipe.component.ComponentRole;
 import dev.latvian.mods.kubejs.recipe.component.RecipeComponent;
 import net.dries007.tfc.common.recipes.ingredients.FluidIngredient;
@@ -38,7 +39,21 @@ public class FluidIngredientComponent implements RecipeComponent<FluidIngredient
         return IngredientHelpers.ofFluidIngredient(from);
     }
 
+    @Override
+    public boolean isInput(RecipeJS recipe, FluidIngredient value, ReplacementMatch match) {
+        final FluidIngredient matchIng = IngredientHelpers.ofFluidIngredient(match);
+        if (!matchIng.entries().isEmpty()) {
+            return matchIng.entries().stream().anyMatch(e -> value.all().anyMatch(e));
+        }
+        return false;
+    }
+
     public static class FluidStackIngredientComponent implements RecipeComponent<FluidStackIngredient> {
+
+        @Override
+        public ComponentRole role() {
+            return ComponentRole.INPUT;
+        }
 
         @Override
         public String componentType() {
@@ -58,6 +73,11 @@ public class FluidIngredientComponent implements RecipeComponent<FluidIngredient
         @Override
         public FluidStackIngredient read(RecipeJS recipe, Object from) {
             return IngredientHelpers.ofFluidStackIngredient(from);
+        }
+
+        @Override
+        public boolean isInput(RecipeJS recipe, FluidStackIngredient value, ReplacementMatch match) {
+            return INGREDIENT.isInput(recipe, value.ingredient(), match);
         }
     }
 }