+++ src/main/java/com/notenoughmail/kubejs_tfc/util/implementation/commands/KubeJSTFCCommands.java
@@ -4,7 +4,6 @@ import com.mojang.brigadier.arguments.DoubleArgumentType;
 import com.mojang.brigadier.arguments.IntegerArgumentType;
 import com.mojang.brigadier.arguments.StringArgumentType;
 import com.mojang.brigadier.context.CommandContext;
-import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import com.notenoughmail.kubejs_tfc.KubeJSTFC;
 import com.notenoughmail.kubejs_tfc.util.implementation.bindings.MiscBindings;
 import com.notenoughmail.kubejs_tfc.util.implementation.mixin.accessor.DataManagerAccessor;
@@ -15,8 +14,6 @@ import net.dries007.tfc.util.DataManager;
 import net.dries007.tfc.world.ChunkGeneratorExtension;
 import net.dries007.tfc.world.chunkdata.ChunkData;
 import net.dries007.tfc.world.chunkdata.LerpFloatLayer;
-import net.dries007.tfc.world.noise.Noise2D;
-import net.dries007.tfc.world.noise.Noise3D;
 import net.dries007.tfc.world.settings.RockLayerSettings;
 import net.dries007.tfc.world.settings.Settings;
 import net.minecraft.ChatFormatting;
@@ -31,8 +28,6 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.util.Mth;
 import net.minecraft.world.level.ChunkPos;
-import net.minecraft.world.level.block.Block;
-import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.levelgen.structure.BoundingBox;
 import net.minecraftforge.event.RegisterCommandsEvent;
 
@@ -114,7 +109,7 @@ public class KubeJSTFCCommands {
                                                         .then(argument("output_range", Range.arg())
                                                                 .then(argument("noise", StringArgumentType.greedyString())
                                                                         .suggests((ctx, builder) -> SharedSuggestionProvider.suggest(MiscBindings.INSTANCE.inspect2DNoise.get().keySet(), builder))
-                                                                        .executes(KubeJSTFCCommands::inspectNoise2D)
+                                                                        .executes(NoiseInspection::inspectNoise2D)
                                                                 )
                                                         )
                                                 )
@@ -134,9 +129,9 @@ public class KubeJSTFCCommands {
                                                                                         .map(s -> s.contains(" ") ? "\"" + s + "\"" : s),
                                                                                 builder
                                                                         ))
-                                                                        .executes(KubeJSTFCCommands::inspectNoise3D)
+                                                                        .executes(NoiseInspection::inspectNoise3D)
                                                                         .then(argument("y_input", DoubleArgumentType.doubleArg())
-                                                                                .executes(KubeJSTFCCommands::inspectNoise3DAtHeight)
+                                                                                .executes(NoiseInspection::inspectNoise3DAtHeight)
                                                                         )
                                                                 )
                                                         )
@@ -354,196 +349,16 @@ public class KubeJSTFCCommands {
         return 1;
     }
 
-    private static int inspectNoise2D(CommandContext<CommandSourceStack> ctx) throws CommandSyntaxException {
-        if (KubeJSTFC.debug) {
-            final String noiseId = StringArgumentType.getString(ctx, "noise");
-            final Noise2D noise = MiscBindings.INSTANCE.inspect2DNoise.get().get(noiseId);
-            if (noise != null) {
-                final BlockPos from = BlockPosArgument.getLoadedBlockPos(ctx, "from"), to = BlockPosArgument.getLoadedBlockPos(ctx, "to");
-
-                final int minY = Math.min(from.getY(), to.getY()), maxY = Math.max(from.getY(), to.getY());
-                final int minX = Math.min(from.getX(), to.getX()), minZ = Math.min(from.getZ(), to.getZ());
-                final int horizontalRange = Math.min(Math.max(from.getX(), to.getX()) - minX, Math.max(from.getZ(), to.getZ()) - minZ);
-                final int verticalRange = maxY - minY;
-
-                if (verticalRange < 2) {
-                    return failMsg("Too short to properly display noise. Please increase the y-range", ctx);
-                }
-
-                final Range rangeIn = Range.get("input_range", ctx), rangeOut = Range.get("output_range", ctx);
-                final double inStep = rangeIn.step(horizontalRange);
-
-                final ServerLevel level = ctx.getSource().getLevel();
-                final BlockPos.MutableBlockPos cursor = new BlockPos.MutableBlockPos();
-                for (int x = 0 ; x < horizontalRange ; x++) {
-                    cursor.setX(x + minX);
-                    for (int z = 0 ; z < horizontalRange ; z++) {
-                        cursor.setZ(z + minZ);
-                        final double noiseVal = noise.noise(
-                                rangeIn.min() + (inStep * x),
-                                rangeIn.min() + (inStep * z)
-                        );
-                        final int noiseY = (int) Mth.map(noiseVal, rangeOut.min(), rangeOut.max(), 0, verticalRange - 1);
-                        boolean placedGlass = false;
-                        for (int y = 0 ; y < verticalRange ; y++) {
-                            cursor.setY(y + minY);
-                            if (y == noiseY) {
-                                level.setBlockAndUpdate(cursor, Blocks.WHITE_STAINED_GLASS.defaultBlockState());
-                                placedGlass = true;
-                            } else {
-                                level.setBlockAndUpdate(cursor, Blocks.AIR.defaultBlockState());
-                            }
-                        }
-                        if (!placedGlass) {
-                            if (noiseY < verticalRange) {
-                                cursor.setY(minY);
-                            }
-                            level.setBlockAndUpdate(cursor, Blocks.RED_STAINED_GLASS.defaultBlockState());
-                        }
-                    }
-                }
-                sysMsg("Displayed noise '%s' in-world".formatted(noiseId), ctx);
-                return 1;
-            } else {
-                return failMsg("Unregistered noise '%s'".formatted(noiseId), ctx);
-            }
-        } else {
-            return failMsg("Debug mode is not enabled, so no noises were ever registered for inspection", ctx);
-        }
-    }
-
-    // Notionally: white --bright--> grays --dark--> black
-    private static final Block[] GRADIENT_3D = {
-            Blocks.WHITE_STAINED_GLASS,
-            Blocks.PINK_STAINED_GLASS,
-            Blocks.RED_STAINED_GLASS,
-            Blocks.ORANGE_STAINED_GLASS,
-            Blocks.YELLOW_STAINED_GLASS,
-            Blocks.LIME_STAINED_GLASS,
-            Blocks.LIGHT_BLUE_STAINED_GLASS,
-            Blocks.LIGHT_GRAY_STAINED_GLASS,
-            Blocks.GRAY_STAINED_GLASS,
-            Blocks.CYAN_STAINED_GLASS,
-            Blocks.GREEN_STAINED_GLASS,
-            Blocks.BLUE_STAINED_GLASS,
-            Blocks.MAGENTA_STAINED_GLASS,
-            Blocks.PURPLE_STAINED_GLASS,
-            Blocks.BROWN_STAINED_GLASS,
-            Blocks.BLACK_STAINED_GLASS
-    };
-
-    /**
-     * Gets the appropriate block in the gradient.
-     * <p>
-     * Accepts any value, but only values in the range [{@code 0}, {@code 15}] are part of the gradient.
-     * Values outside that range will return air.
-     */
-    public static Block getGradient(int val) {
-        if (val < 0 || val > 15) {
-            return Blocks.AIR;
-        }
-        return GRADIENT_3D[val];
-    }
-
-    private static int inspectNoise3D(CommandContext<CommandSourceStack> ctx) throws CommandSyntaxException {
-        if (KubeJSTFC.debug) {
-            final String noiseId = StringArgumentType.getString(ctx, "noise");
-            final Noise3D noise = MiscBindings.INSTANCE.inspect3DNoise.get().get(noiseId);
-            if (noise != null) {
-                final BlockPos from = BlockPosArgument.getLoadedBlockPos(ctx, "from"), to = BlockPosArgument.getLoadedBlockPos(ctx, "to");
-
-                final int minY = Math.min(from.getY(), to.getY()), minX = Math.min(from.getX(), to.getX()), minZ = Math.min(from.getZ(), to.getZ());
-                final int volumetricRange = Math.min(
-                        Math.min(
-                                Math.max(from.getY(), to.getY()) - minY,
-                                Math.max(from.getX(), to.getX()) - minX
-                        ),
-                        Math.max(from.getZ(), to.getZ()) - minZ
-                );
-
-                final Range rangeIn = Range.get("input_range", ctx), rangeOut = Range.get("output_range", ctx);
-                final double inStep = rangeIn.step(volumetricRange);
-
-                final ServerLevel level = ctx.getSource().getLevel();
-                final BlockPos.MutableBlockPos cursor = new BlockPos.MutableBlockPos();
-                for (int x = 0 ; x < volumetricRange ; x++) {
-                    cursor.setX(x + minX);
-                    for (int y = 0 ; y < volumetricRange ; y++) {
-                        cursor.setY(y + minY);
-                        for (int z = 0 ; z < volumetricRange ; z++) {
-                            cursor.setZ(z + minZ);
-                            final double noiseVal = noise.noise(
-                                    rangeIn.min() + (inStep * x),
-                                    rangeIn.min() + (inStep * y),
-                                    rangeIn.min() + (inStep * z)
-                            );
-                            final Block block = getGradient((int) Math.round(Mth.map(
-                                    noiseVal,
-                                    rangeOut.min(),
-                                    rangeOut.max(),
-                                    0,
-                                    15
-                            )));
-                            level.setBlockAndUpdate(cursor, block.defaultBlockState());
-                        }
-                    }
-                }
-
-                sysMsg("Displayed noise '%s' in-world".formatted(noiseId), ctx);
-                return 1;
-            } else {
-                return failMsg("Unregistered noise '%s'".formatted(noiseId), ctx);
-            }
-        } else {
-            return failMsg("Debug mode is not enabled, so no noises were ever registered for inspection", ctx);
-        }
-    }
-
-    private static int inspectNoise3DAtHeight(CommandContext<CommandSourceStack> ctx) {
-        final String noiseName = StringArgumentType.getString(ctx, "noise");
-        final Noise3D noise = MiscBindings.INSTANCE.inspect3DNoise.get().get(noiseName);
-        if (noise == null) {
-            return failMsg("Unregistered noise '%s'".formatted(noiseName), ctx);
-        } else {
-            final double y = DoubleArgumentType.getDouble(ctx, "y_input");
-
-            final String name2D = noiseName + " at %.1f".formatted(y);
-            MiscBindings.INSTANCE.register2DNoiseForInspection(name2D, (x, z) -> noise.noise(x, y, z));
-
-            final Range rangeIn = Range.get("input_range", ctx), rangeOut = Range.get("output_range", ctx);
-            final BlockPos fromPos = BlockPosArgument.getBlockPos(ctx, "from"), toPos = BlockPosArgument.getBlockPos(ctx, "to");
-
-            final int ret = ctx.getSource().getServer().getCommands().performPrefixedCommand(
-                    ctx.getSource(),
-                    "/kubejs_tfc inspect_2d_noise %d %d %d %d %d %d %f %f %f %f %s".formatted(
-                            fromPos.getX(),
-                            fromPos.getY(),
-                            fromPos.getZ(),
-                            toPos.getX(),
-                            toPos.getY(),
-                            toPos.getZ(),
-                            rangeIn.min(),
-                            rangeIn.max(),
-                            rangeOut.min(),
-                            rangeOut.max(),
-                            name2D
-                    ) // I can't be bothered to generalize the method
-            );
-            MiscBindings.INSTANCE.inspect2DNoise.get().remove(name2D);
-            return ret;
-        }
-    }
-
-    private static int failMsg(String msg, CommandContext<CommandSourceStack> ctx) {
+    static int failMsg(String msg, CommandContext<CommandSourceStack> ctx) {
         ctx.getSource().sendFailure(Component.literal(msg));
         return 0;
     }
 
-    private static void sysMsg(String msg, CommandContext<CommandSourceStack> ctx) {
+    static void sysMsg(String msg, CommandContext<CommandSourceStack> ctx) {
         ctx.getSource().sendSystemMessage(Component.literal(msg));
     }
 
-    private static void sysMsg(Component cmp, CommandContext<CommandSourceStack> ctx) {
+    static void sysMsg(Component cmp, CommandContext<CommandSourceStack> ctx) {
         ctx.getSource().sendSystemMessage(cmp);
     }